requirements_streamlit.txt
# Requirements for Streamlit Proof System Visualizers

# Core Streamlit
streamlit>=1.28.0

# Visualization
plotly>=5.17.0
networkx>=3.1
matplotlib>=3.7.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Existing requirements
crewai>=0.30.0
langchain>=0.1.0
langchain-openai>=0.0.8
faiss-cpu==1.10.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2

# arXiv integration
arxiv>=2.3.0

# Google Generative AI for markdown reports (optional)
google-generativeai>=0.3.0

# Additional for real-time features
streamlit-autorefresh>=0.0.1
watchdog>=3.0.0